+++ core/src/main/java/org/nervos/appchain/tx/Contract.java
@@ -242,8 +242,8 @@ public abstract class Contract extends ManagedTransaction {
 
     // adapt to cita
     TransactionReceipt executeTransaction(
-            String data, BigInteger quota, BigInteger nonce, BigInteger validUntilBlock,
-            BigInteger version , int chainId, String value)
+            String data, long quota, BigInteger nonce, long validUntilBlock,
+            int version , int chainId, String value)
             throws TransactionException, IOException {
         return sendAdaptToCita(
                 contractAddress, data, quota, nonce, validUntilBlock, version, chainId, value);
@@ -278,8 +278,8 @@ public abstract class Contract extends ManagedTransaction {
     }
 
     protected RemoteCall<TransactionReceipt> executeRemoteCallTransaction(
-            Function function, BigInteger quota, BigInteger nonce,
-            BigInteger validUntilBlock, BigInteger version,
+            Function function, long quota, BigInteger nonce,
+            long validUntilBlock, int version,
             int chainId, String value) {
         return new RemoteCall<>(
                 () -> executeTransaction(
@@ -305,8 +305,8 @@ public abstract class Contract extends ManagedTransaction {
 
     private static <T extends Contract> T create(
             T contract, String binary, String encodedConstructor,
-            BigInteger quota, BigInteger nonce, BigInteger validUntilBlock,
-            BigInteger version, int chainId, String value)
+            long quota, BigInteger nonce, long validUntilBlock,
+            int version, int chainId, String value)
             throws IOException, TransactionException {
         TransactionReceipt transactionReceipt =
                 contract.executeTransaction(
@@ -376,8 +376,8 @@ public abstract class Contract extends ManagedTransaction {
     protected static <T extends Contract> T deploy(
             Class<T> type,
             Nervosj nervosj, TransactionManager transactionManager,
-            BigInteger quota, BigInteger nonce, BigInteger validUntilBlock,
-            BigInteger version, String binary, int chainId,
+            long quota, BigInteger nonce, long validUntilBlock,
+            int version, String binary, int chainId,
             String value, String encodedConstructor)
             throws IOException, TransactionException {
 
@@ -401,8 +401,8 @@ public abstract class Contract extends ManagedTransaction {
     protected static <T extends Contract> RemoteCall<T>
                     deployRemoteCall(
             Class<T> type, Nervosj nervosj, TransactionManager transactionManager,
-            BigInteger quota, BigInteger nonce, BigInteger validUntilBlock,
-            BigInteger version, int chainId, String value,
+            long quota, BigInteger nonce, long validUntilBlock,
+            int version, int chainId, String value,
             String binary, String encodedConstructor) {
         return new RemoteCall<>(() -> deploy(
                 type, nervosj, transactionManager, quota, nonce, validUntilBlock,
